refactor: clean the code base and make some of code generic
