+++ packages/server/shared/src/lib/system/system-prop.ts
@@ -53,6 +53,7 @@ export enum AppSystemProp {
     TRIGGER_FAILURES_THRESHOLD = 'TRIGGER_FAILURES_THRESHOLD',
     MAX_CONCURRENT_JOBS_PER_PROJECT = 'MAX_CONCURRENT_JOBS_PER_PROJECT',
     PROJECT_RATE_LIMITER_ENABLED = 'PROJECT_RATE_LIMITER_ENABLED',
+    DEV_PIECES = 'DEV_PIECES',
 
 
     // ENTERPRISE ONLY
