+++ packages/server/worker/src/lib/utils/cache-handler.ts
@@ -4,11 +4,7 @@ import { fileExists, memoryLock, SharedSystemProp, system, threadSafeMkdir } fro
 import {  ApEnvironment, isNil } from '@activepieces/shared'
 import writeFileAtomic from 'write-file-atomic'
 
-export enum CacheState {
-    READY = 'READY',
-    PENDING = 'PENDING',
-}
-type CacheMap = Record<string, CacheState | string>
+type CacheMap = Record<string, string>
 const isDev = system.getOrThrow(SharedSystemProp.ENVIRONMENT) === ApEnvironment.DEVELOPMENT
 
 const cachePath = (folderPath: string): string => join(folderPath, 'cache.json')
@@ -29,9 +25,9 @@ const getCache = async (folderPath: string): Promise<CacheMap> => {
 
 export const cacheHandler = (folderPath: string) => {
     return {
-        async cacheCheckState(cacheAlias: string): Promise<CacheState | string | undefined> {
+        async cacheCheckState(cacheAlias: string): Promise<string | undefined> {
             if (isDev) {
-                return CacheState.PENDING
+                return undefined
             }
             const lock = await memoryLock.acquire('cache_' + cacheAlias)
             try {
@@ -42,7 +38,7 @@ export const cacheHandler = (folderPath: string) => {
                 await lock.release()
             }
         },
-        async setCache(cacheAlias: string, state: CacheState | string): Promise<void> {
+        async setCache(cacheAlias: string, state: string): Promise<void> {
             const lock = await memoryLock.acquire('cache_' + cacheAlias)
             try {
                 const cache = await getCache(folderPath)
